pi-image-vending/requirements.txt
RPi.Gpio==0.7.1
requests==2.28.2
